06_PS_VAE/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
einops>=0.7.0
numpy>=1.24.0
pillow>=9.0.0
tqdm>=4.65.0
wandb>=0.15.0
omegaconf>=2.3.0
accelerate>=0.20.0
transformers>=4.30.0
diffusers>=0.25.0
lpips>=0.1.4
pytorch-fid>=0.3.0

